apsdfo

+++ app/Http/Controllers/CatalogoPiezasController.php
@@ -35,22 +35,27 @@ class CatalogoPiezasController extends Controller
     public function store(Request $request)
     {
 
-        $this->validate($request, [
-            'nombre_pieza'=> 'required',
-            'descripcion_pieza'=> 'required',
-            'url_imagen_pieza'=> 'required|url|active_url',
+        if(Auth::guard('admin')->check()){
+            $this->validate($request, [
+                'nombre_pieza'=> 'required',
+                'descripcion_pieza'=> 'required',
+                'url_imagen_pieza'=> 'required|url|active_url',
 
-        ]);
+            ]);
 
-        $pieza = new CatalogoPiezas;
-        $pieza->nombre_pieza = $request->input('nombre_pieza');
-        $pieza->descripcion_pieza = $request->input('descripcion_pieza');
-        $pieza->url_img_pieza = $request->input('url_imagen_pieza');
-    
-        $pieza->save();
-    
-    
-        return redirect('/admin/catalogo')->with('success','Pieza añadida exitosamente.');
+            $pieza = new CatalogoPiezas;
+            $pieza->nombre_pieza = $request->input('nombre_pieza');
+            $pieza->descripcion_pieza = $request->input('descripcion_pieza');
+            $pieza->url_img_pieza = $request->input('url_imagen_pieza');
+        
+            $pieza->save();
+        
+        
+            return redirect('/admin/catalogo')->with('success','Pieza añadida exitosamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function show($id)
@@ -61,35 +66,50 @@ class CatalogoPiezasController extends Controller
 
     public function edit($id)
     {
-        $pieza = CatalogoPiezas::find($id);
-        return view('editarpieza')->with('pieza', $pieza); // Cambiar cuando se vaya a motrar algo
+        if(Auth::guard('admin')->check()){
+            $pieza = CatalogoPiezas::find($id);
+            return view('editarpieza')->with('pieza', $pieza); // Cambiar cuando se vaya a motrar algo
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'nombre_pieza'=> 'required',
-            'descripcion_pieza'=> 'required',
-            'url_imagen_pieza'=> 'required|url|active_url',
+        if(Auth::guard('admin')->check()){
+            $this->validate($request, [
+                'nombre_pieza'=> 'required',
+                'descripcion_pieza'=> 'required',
+                'url_imagen_pieza'=> 'required|url|active_url',
 
-        ]);
+            ]);
 
-        $pieza = CatalogoPiezas::find($id);
-        $pieza->nombre_pieza = $request->input('nombre_pieza');
-        $pieza->descripcion_pieza = $request->input('descripcion_pieza');
-        $pieza->url_img_pieza = $request->input('url_imagen_pieza');
-    
-        $pieza->save();
-    
-    
-        return redirect('/admin/catalogo')->with('success','Pieza actualizada exitosamente.');
+            $pieza = CatalogoPiezas::find($id);
+            $pieza->nombre_pieza = $request->input('nombre_pieza');
+            $pieza->descripcion_pieza = $request->input('descripcion_pieza');
+            $pieza->url_img_pieza = $request->input('url_imagen_pieza');
+        
+            $pieza->save();
+        
+        
+            return redirect('/admin/catalogo')->with('success','Pieza actualizada exitosamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function destroy($id)
     {
-        $pieza = CatalogoPiezas::find($id);
-        $pieza->delete();
-        
-        return redirect('/admin/catalogo')->with('success', 'La pieza se ha eliminado correctamente.');
+        if(Auth::guard('admin')->check()){
+            $pieza = CatalogoPiezas::find($id);
+            $pieza->delete();
+            
+            return redirect('/admin/catalogo')->with('success', 'La pieza se ha eliminado correctamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 }
